requirements.txt
blinker==1.7.0
cffi==1.16.0
click==8.1.7
cryptography==41.0.7
Flask==3.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jwt==1.3.1
MarkupSafe==2.1.3
psycopg==3.1.14
psycopg-binary==3.1.14
pycparser==2.21
python-dotenv==1.0.0
typing_extensions==4.8.0
Werkzeug==3.0.1
